+++ app/src/main/java/com/ganaljigi/kubf/data/remote/repository/HomeRepository.kt
@@ -1,7 +1,7 @@
 package com.ganaljigi.kubf.data.remote.repository
 
-import com.ganaljigi.kubf.data.remote.response.HomeResponseDto
-import com.ganaljigi.kubf.data.remote.response.HomeSignificantResponseDto
+import com.ganaljigi.kubf.data.remote.response.home.HomeResponseDto
+import com.ganaljigi.kubf.data.remote.response.home.HomeSignificantResponseDto
 
 interface HomeRepository {
     suspend fun getHomeData(): Result<HomeResponseDto>
