+++ app/src/main/java/com/ganaljigi/kubf/data/remote/service/HomeService.kt
@@ -1,8 +1,8 @@
 package com.ganaljigi.kubf.data.remote.service
 
 import com.ganaljigi.kubf.data.remote.base.BaseResponse
-import com.ganaljigi.kubf.data.remote.response.HomeResponseDto
-import com.ganaljigi.kubf.data.remote.response.HomeSignificantResponseDto
+import com.ganaljigi.kubf.data.remote.response.home.HomeResponseDto
+import com.ganaljigi.kubf.data.remote.response.home.HomeSignificantResponseDto
 import retrofit2.http.GET
 import retrofit2.http.Path
